Split trendline indicator into several files.
