+++ js/masters/indicators/trendline.src.js
@@ -11,4 +11,4 @@
  * License: www.highcharts.com/license
  */
 'use strict';
-import '../../Stock/Indicators/TrendLineIndicator.js';
+import '../../Stock/Indicators/TrendLine/TrendLineIndicator.js';
