+++ ts/Stock/Indicators/TrendLine/TrendLinePoint.d.ts
@@ -1,0 +1,33 @@
+/* *
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import TrendLineIndicator from './TrendLineIndicator';
+import type SMAPoint from '../SMA/SMAPoint';
+
+/* *
+*
+*  Class
+*
+* */
+
+declare class TrendLinePoint extends SMAPoint {
+    public series: TrendLineIndicator;
+}
+
+/* *
+*
+*  Default Export
+*
+* */
+
+export default TrendLinePoint;
